backend/requirements.txt
# --- Core API / DB ---
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.32
# psycopg2-binary==2.9.9
psycopg[binary]>=3.2.4
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
pydantic==2.8.2

# --- Ingestion ---
pandas==2.2.2
pdfplumber==0.11.4
rapidfuzz==3.9.6
python-multipart==0.0.9
aiofiles==24.1.0

# --- Optional (keep commented until used) ---
# alembic==1.13.2
# pytest==8.2.1
# httpx==0.27.0
# pytesseract==0.3.13
# Pillow==10.4.0
# scikit-learn==1.5.1
# joblib==1.4.2
# sentence-transformers==3.0.1
# camelot-py==0.11.0
# tabula-py==2.9.3
